requirements.txt
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pathlib2>=2.3.0; python_version < "3.4"